apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
yfinance==0.2.40
pydantic==2.7.4
pydantic-settings==2.3.0
httpx==0.27.0
ruff==0.4.10
pytest==8.2.2
pytest-asyncio==0.23.7